Getting the plugin to register on iOS in a Cordova 3.0 environment
The signature of all the methods callable from JavaScript had to change
from:
-- (void)register:(NSMutableArray *)arguments
withDict:(NSMutableDictionary *)options;
to
+- (void)register:(CDVInvokedUrlCommand*)command;

The old format was officially obsoleted in in Cordova 3.0, so there is
no way around using the old one.

'arguments' and 'options' needed to use the new 'command' object.

With this new format the plugin result interface was also broken, so it
was changed to the new documented format.

This code is _not_ well tested at this point, but I wanted to get it up
for anyone having issues migrating to 3.0.