Android option `updateNotifications`
Upon receiving a notification message, the current behavior will update
the existing notification message with the new content. In some cases,
this is not the desired behavior, which is where `updateNotifications`
comes in. By setting it to to `false` (from the default `true`), each
new message will produce a new notification message.