Retrieve APNs environment
Read the embedded.mobileprovision file and retrieve the
'aps-environement' entitlement to determine whether we're using the
development (sandbox) environment or the production environment.

Add "apsEnvironment" (string), "apsIsProduction" (boolean), and
"apsIsDevelopment"  (boolean)keys to the object passed to the
"registration" event handler.

One can then pass the information to the backend server to use the
correct environment (certificate, key, host).