Restore custom Android permission declaration
Custom Android permission was introduced according to
"Adding custom permission for Push handler Activity #1362".

However, it was found that the permission declaration is
missing after "Merge v2.0.x into master (#1736)".

There is no much sense in requesting a signature level
permission that is not declared in the same application
manifest.

Fix the merge conflict and restore the custom permission
declaration.